refactor: 숫자리스트로 바꾸기 전에 0이상 숫자인지 먼저 확인
